add module folder to seperate the js files
